+++ src/rule_providers.ts
@@ -98,4 +98,12 @@ export const ruleProviders: Record<string, RuleProvider> = {
         url: `${CDN_URL}/gh/powerfullz/override-rules@master/ruleset/Weibo.list`,
         path: "./ruleset/Weibo.list",
     },
+    GFWList: {
+        type: "http",
+        behavior: "domain",
+        format: "yaml",
+        interval: 86400,
+        url: "https://cdn.jsdelivr.net/gh/Loyalsoldier/clash-rules@release/gfw.txt",
+        path: "./ruleset/GFWList.yaml",
+    },
 };